requirements.txt
concrete-ml==1.9.0
matplotlib==3.10.3
numpy==1.26.4
opencv-python==4.11.0.86
scikit-learn==1.5.0
tqdm==4.67.1
python_version >=3.10.13
